requirements.txt
fastapi>=0.110
langchain-core>=0.2
langchain-openai>=0.1
langchain>=0.2
langgraph>=0.0.50
langsmith>=0.1
pinecone-client>=4.1
pydantic>=2.6
python-dotenv>=1.0
uvicorn[standard]>=0.30
